+++ internal/webhook/proxmoxmachine_webhook_test.go
@@ -44,7 +44,7 @@ var _ = Describe("Controller Test", func() {
 
 		It("should disallow invalid network mtu for additional device", func() {
 			machine := validProxmoxMachine("test-machine")
-			machine.Spec.Network.NetworkDevices[0].MTU = ptr.To(uint16(1000))
+			machine.Spec.Network.NetworkDevices[0].MTU = ptr.To(int32(1000))
 			g.Expect(k8sClient.Create(testEnv.GetContext(), &machine)).To(MatchError(ContainSubstring("mtu must be at least 1280 or 1, but was 1000")))
 		})
 
@@ -55,13 +55,13 @@ var _ = Describe("Controller Test", func() {
 
 		It("should disallow invalid network vlan for additional device", func() {
 			machine := validProxmoxMachine("test-machine")
-			machine.Spec.Network.NetworkDevices[0].VLAN = ptr.To(uint16(0))
+			machine.Spec.Network.NetworkDevices[0].VLAN = ptr.To(int32(0))
 			g.Expect(k8sClient.Create(testEnv.GetContext(), &machine)).To(MatchError(ContainSubstring("greater than or equal to 1")))
 		})
 
 		It("should disallow invalid link mtu for additional device", func() {
 			machine := validProxmoxMachine("test-machine")
-			machine.Spec.Network.NetworkDevices[0].LinkMTU = ptr.To(uint16(1000))
+			machine.Spec.Network.NetworkDevices[0].LinkMTU = ptr.To(int32(1000))
 			g.Expect(k8sClient.Create(testEnv.GetContext(), &machine)).To(MatchError(ContainSubstring("mtu must be at least 1280, but was 1000")))
 		})
 
@@ -85,10 +85,10 @@ var _ = Describe("Controller Test", func() {
 			g.Expect(k8sClient.Create(testEnv.GetContext(), &machine)).To(Succeed())
 
 			g.Expect(k8sClient.Get(testEnv.GetContext(), client.ObjectKeyFromObject(&machine), &machine)).To(Succeed())
-			machine.Spec.Network.NetworkDevices[0].MTU = ptr.To(uint16(50))
+			machine.Spec.Network.NetworkDevices[0].MTU = ptr.To(int32(50))
 
 			g.Expect(k8sClient.Update(testEnv.GetContext(), &machine)).To(MatchError(ContainSubstring("spec.network.NetworkDevices[0].mtu: Invalid value")))
-			machine.Spec.Network.NetworkDevices[0].VLAN = ptr.To(uint16(0))
+			machine.Spec.Network.NetworkDevices[0].VLAN = ptr.To(int32(0))
 
 			g.Expect(k8sClient.Update(testEnv.GetContext(), &machine)).To(MatchError(ContainSubstring("spec.network.NetworkDevices[0].vlan: Invalid value")))
 
@@ -137,14 +137,14 @@ func validProxmoxMachine(name string) infrav1.ProxmoxMachine {
 					Name:   "net0",
 					Bridge: "vmbr1",
 					Model:  ptr.To("virtio"),
-					MTU:    ptr.To(uint16(1500)),
-					VLAN:   ptr.To(uint16(100)),
+					MTU:    ptr.To(int32(1500)),
+					VLAN:   ptr.To(int32(100)),
 				}, {
 					Name:   "net1",
 					Bridge: "vmbr2",
 					Model:  ptr.To("virtio"),
-					MTU:    ptr.To(uint16(1500)),
-					VLAN:   ptr.To(uint16(100)),
+					MTU:    ptr.To(int32(1500)),
+					VLAN:   ptr.To(int32(100)),
 					InterfaceConfig: infrav1.InterfaceConfig{
 						IPPoolRef: []corev1.TypedLocalObjectReference{{
 							Name:     "simple-pool",
@@ -180,7 +180,7 @@ func invalidMTUProxmoxMachine(name string) infrav1.ProxmoxMachine {
 		Name:   "net0",
 		Bridge: "vmbr1",
 		Model:  ptr.To("virtio"),
-		MTU:    ptr.To(uint16(50)),
+		MTU:    ptr.To(int32(50)),
 	}}
 	return machine
 }
@@ -191,7 +191,7 @@ func invalidVLANProxmoxMachine(name string) infrav1.ProxmoxMachine {
 		Name:   "net0",
 		Bridge: "vmbr1",
 		Model:  ptr.To("virtio"),
-		VLAN:   ptr.To(uint16(0)),
+		VLAN:   ptr.To(int32(0)),
 	}}
 	return machine
 }
